include player property updates in "changes" segment
Updates for select properties are included in outgoing messages (in
changes.stat_values) whenever their values change.
The implementation of detecting those changes is not particularly
elegant (a boolean flag that is set to true in setter methods, and
reset to false when changes are sent out), but simple enough and it
works within the architectural constraints (the Property instances do
not "know" their "owner").