abort message sending (no error) when no session available
At some point in the player creation process, the GSJS code is
trying to send messages to the player that is just being created.
It is ok to just ignore that attempt in this scenario (and would
require modifying GSJS code to prevent).
In other circumstances though, the same symptom may be indicative of a
problem (e.g. an offline player still in a locations player list), so
log it at INFO level here for now, including a stack trace so we know
where the call is coming from.