don't implicitly mark objects scheduled for unloading as dirty
Fixes trello#21.

When an object is scheduled for unloading at the end of the request,
don't automatically mark it as dirty; it may not even be loaded, in
which case this would cause the object to be loaded unnecessarily. If
there are changes to be persisted, the object must have been marked as
dirty by the standard mechanism (persistence helper proxy).