don't invoke callback twice if the callback itself throws
Problem uncovered while testing the fix for trello#21: the "regular"
callback invocation was inside a try block whose catch clause called
the same callback again in case of errors.
This change restructures the RC.run function so that the try-catch
only covers the request function call, not the part which invokes the
callback. pers.postRequestProc is not expected to throw (async
function); if it still does, the error should be handled by the
Session domain.