add simplistic deployment script
A very simple bash script deployment system. deploy.sh is meant to be
run via a Jenkins job (which sets the necessary environment variables).
It rsyncs the GS package to the target system, and invokes
deploy-remote.sh there to replace the old version with the new one,
and restart the GS service.