don't modify message object when sending
There is at least one case where GSJS code recycles the object passed
as a message for the client for a different purpose (adding buff
details to the player's events list, which eventually writes this
object to persistence). Therefore, the send function must not modify
the object.
The change in Player.send makes cloning the message object in
Location.send redundant.