start server with --expose-gc flag
After rather extensive analysis of what looks like memory leak
symptoms (particularly on the alpha server), and attempts at finding
and fixing possible causes, it turns out that simply running node with
the '--expose-gc' flag apparently fixes all our problems in that
regard.
This really does not make much sense, but for the sake of further
analysis and gaining experience, let's use this by default when starting
up the GS for now.