remove live object cache compaction functionality
This is a partial revert of 7a396dc (revert the actual feature, but
keep some of the utility functions etc). Cache compaction does not
have the desired effect (the GC still does not free memory as intended,
most likely due to dangling references to the released objects, or parts
of them), but instead introduces consistency issues. This is an
inherent problem, as the game objects are not atomic entities - any
part of the code (including GSJS) can simply store its own references
to (parts of) other objects, which cannot be reliably updated when those
objects are released from cache.