actually wrap game objects in proxy wrapper
The call to orProxy.wrap introduced in 0997804 is actually totally
ineffective (since the return value is not used). This change fixes
that, and is quite a bit more complex, since we cannot simply wrap the
objects in the live object cache (that would lead to an obvious
infinite recursion when accessing properties), but have to handle it
in get and create instead.