ensure changes&announcements are sent after request processing
The GSJS code is not written in a way that makes sure announcements
resulting from a request are actually sent out while processing it
(example: in quoin.onPlayerCollision, there is a "broadcastState"
call, but it appears *before* the quoin_got and respective sound anncs
are queued).
This change causes any pending announcements and changes to always be
sent right after the GSJS request processing finishes. It does seem
like a brute force approach to the problem - a more elegant solution
might come up in the future.