fix "not exists in location" error popup
This fixes the "not exists in location, but a delete changes was sent
for it" error when reviving trants with fertilidust or assembling
machines.
It is more workaround than clean solution; the client behaves oddly
when receiving changes for new and deleted items within one message,
and this just prevents such combinations from happening.

When the client receives a message of type ITEMSTACK_STATE, it checks to
see if the itemstack (im.payload.itemstack_tsid) exists in the world
(see do_ITEMSTACK_STATE in NCMsgHandler.as). If it doesn't, then it
processes the changes. Later, it processes the same changes again in
postProcessMessage (NetController.as).