prevent errors when trants are replace with dead ones
When trants die, a new item of type trant_xyz_dead replaces the trant
item. The resulting "onContainerItemAdded" event is sent to all items
in the location, including the just deleted trant items, which fails.
A very similar problem can occur when the item "onPlayerEnter" or
"onPlayerExit" events are processed for all items in a location.