don't send changes for hidden items
Don't consider hidden items (furniture, escrow, trophy bags etc.) when
generating item changes, otherwise these may show up in the client as
white slots or stoot heads in the player inventory.
This happens e.g. when logging in for the first time with a player that
does not have those items yet (in fixtures); the issue did not surface
earlier because the changes used to be sent within a ping reply message
and were thus ignored by the client, which changed in 5a00ded.

Closes #122.