replace non-persistable property values with null
Set game object properties with values that cannot be persisted in
RethinkDB (NaN, Infinity, undefined) to null. RethinkDB does not
currently have a built-in mechanism to handle this itself.
This conversion is performed under the assumption that the GSJS code
never sets those values on purpose.

Closes #133
Fixes trello#83 https://trello.com/c/MulTsxDa