fix changing visit count on location loading screen
The `location.onPlayerEnter` handler was being called too early in the
location move process (during the `relogin_start` request, resp. the
`..._move_end` request for intra-GS moves), causing the location visit
count to be increased and then sent to the client again (as part of the
response) before the loading process was finished.
This change moves the event handler call after the request, so the visit
counter does not change while the location is being loaded; it also fixes
the order of announcements in the chat window in this situation
("OMG! There's a Qurazy Quoin" message comes *after* "Your left X for
Y").