do not modify log message payload object
When supplying a data object to a logging function (such as `log.info`),
that object should not be modified in any way.
The `harmony-reflect` is required here because it needs to be loaded
before `lodash` (otherwise the lodash cloning functions throw errors when
operating on ES6 proxies).

Closes #140
Fixes trello#122 https://trello.com/c/DFakmeeZ