Prevent timers firing on objects without a container
Prevent worker crashes due to an uncaught error when a timer fires on
an object whose RQ has been closed (typically because the respective
location has been unloaded). This is merely a workaround for a problem
that should not occur in the first place, but the underlying issue is
not clear yet).