Clean up obsolete instance group references
GSJS code removes empty instances after a while by deleting the
respective loc/geo/group objects, but does not remove the group
reference from the template's instance list DC. Do this when an
instance group is deleted, otherwise errors occur when the next
player attempts to move to the location.
This is most probably not the "right" way to do it; it looks like the
persistence layer is meant to return a falsy value in place of deleted/
unavailable object references, which is then explicitly handled in
GSJS (c.f. the `if (!tmp)` block in
`locations/inc_instances.instances_next_instance` in this particular
case).