Apply GSJS base classes in the right order for players
Fix the GSJS prototype composition order for the player class. It used
to add the `item` and `bag` functions to the `player` prototype, which
caused at least two functions in player objects to be incorrectly
referring to their bag counterparts (`distanceFromPlayer` and
`getLocation`). This resulted in being unable to drop the Craftybot.