Prevent persistence errors during shutdown
When the currently loaded game objects are persisted on GS shutdown,
the serialization process might attempt to access objects (to generate
objrefs) that have already been unloaded, causing errors, since objects
may not be loaded during shutdown.
To circumvent this problem, don't clear the proxy object from the cache
when unloading objects. The proxy is sufficient for generating objrefs,
even when the referred-to object is already gone.
The downside to this approach is that proxy objects are never unloaded,
but since these objects are relatively tiny, this seems acceptable for
now.