Actually let cubimals go when they are set free
The `setFreeEnd` NPC movement callback in the cubimal GSJS classes does
not return `true` for full change data, so the compact changes sent to
the client do not convey the information that the item has in fact been
deleted. Include the necessary `count` field in the compact change data
in such cases.

Fixes trello#123 https://trello.com/c/3FTXOPD4