Make sure timers can always access their respective object
Ensure timer calls always have a reference to their corresponding game
object (and not just a proxy). This is breaking the proxy abstraction
layer, but it is the only way to let the timer reliably check if the
object is still "live". The other (equally messy) option would be some
sort of `isLoaded` call on the persistence layer.
The `stale` field needs to be checked both before enqueing the called
function in the RQ, and before finally executing it, since in both
cases the object may have been unloaded in the meantime.

Related: 43613a8, 30eaf5a