Index bag slots
* `Bag.getSlot` currently goes through all of the stacks in `items`
until it finds the correct `x` or slot. This results in performance
issues with bags that have a large number of items such as SDBs. To fix
this, create an index using `Bag.getSlots` and keep it updated when an
items container is changed.

Fixes https://trello.com/c/i7S7x7Nm